fix(cli): share Go toolchain filtering across ELF and dSYM source bundling
